requirements.txt
opencv-python==4.5.4.60
pyyaml
json_tricks
yacs
scikit-learn
pandas
timm==0.4.12
numpy==1.21.2
einops==0.4.1
fvcore
transformers==4.19.2
sentencepiece
ftfy
regex
nltk
vision-datasets==0.2.2
diffdist
pyarrow
cityscapesscripts
shapely
scikit-image
mup
kornia==0.6.4
torchmetrics==0.6.0
blobfile==1.2.7
scipy==1.7.3
pillow==9.5.0
sentence-transformers
protobuf=3.20.0
spacy
nltk
word2number
simple-lama-inpainting